- La view per la visualizzazione ora è la stessa dell'edit (impianto) con la differenza che l'utente che non può modificare l'impianto avrà tutti i campi disabilitati e potrà navigare tra gli step liberamente; - M1/M2 potranno essere modificati solo da chi ha il permesso di farlo, mentre chi non lo ha può leggere i dati al suo interno direttamente dalla lista degli M1/M2; - Altri fix minori;

+++ app/Http/Livewire/Systems/EditMone.php
@@ -20,15 +20,19 @@
 		}
 
 		public function delete() {
-			$this->m_one->delete();
-			$this->emit('m-deleted');
+			if (auth()->user()->can('mone_delete')) {
+				$this->m_one->delete();
+				$this->emit('m-deleted');
+			}
 			$this->closeModal();
 		}
 
 		public function save() {
-			$validated = $this->validate();
-			$this->m_one->update($validated['m_one']);
-			$this->emit('m-updated');
+			if (auth()->user()->can('mone_update')) {
+				$validated = $this->validate();
+				$this->m_one->update($validated['m_one']);
+				$this->emit('m-updated');
+			}
 			$this->closeModal();
 		}
